Move constants into msgraph.cli.core
