+++ msgraph/cli/__init__.py
@@ -2,3 +2,5 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 # ------------------------------------
+
+from .core.profile import read_profile, write_profile
