+++ msgraph/cli/core/constants.py
@@ -7,6 +7,7 @@ from pathlib import Path
 from os import path
 
 AUTH_RECORD_LOCATION = path.join(Path.home(), '.mg', 'record.txt')
+PROFILE_LOCATION = path.join(Path.home(), '.mg', 'profile.json')
 CLIENT_ID = 'f7218512-c727-4138-9fb9-a0fe2500650c'
 
 CLI_COMMON_KWARGS = [
@@ -24,3 +25,26 @@ EXCLUDED_PARAMS = [
     'self', 'raw', 'polling', 'custom_headers', 'operation_config', 'content_version', 'kwargs',
     'client', 'no_wait'
 ]
+
+DEFAULT_CLOUDS = {
+    'PUBLIC': {
+        'endpoint': 'https://graph.microsoft.com',
+        'authority': 'https://login.microsoftonline.com'
+    },
+    'US_GOV_L4': {
+        'endpoint': 'https://graph.microsoft.us',
+        'authority': 'https://login.microsoftonline.us'
+    },
+    'US_GOV_L5': {
+        'endpoint': 'https://dod-graph.microsoft.us',
+        'authority': 'https://login.microsoftonline.us'
+    },
+    'GERMANY': {
+        'endpoint': 'https://graph.microsoft.de',
+        'authority': 'https://login.microsoftonline.de'
+    },
+    'CHINA': {
+        'endpoint': 'https://microsoftgraph.chinacloudapi.cn',
+        'authority': 'https://login.chinacloudapi.cn'
+    }
+}
