+++ msgraph/command_modules/profile/custom.py
@@ -3,17 +3,10 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
-
-from os import path
-from pathlib import Path
 import json
 from knack.prompting import prompt_choice_list
-from msgraph.cli.core.exceptions import CLIException
-
-from .constants import CLOUDS
-
-#TODO: Move this variable to constants.py
-PROFILE_LOCATION = path.join(Path.home(), '.mg', 'profile.json')
+from msgraph.cli.core.constants import DEFAULT_CLOUDS
+from msgraph.cli import read_profile, write_profile
 
 
 def select_cloud():
@@ -22,19 +15,20 @@ def select_cloud():
     user_defined_clouds = profile.get('user_defined_clouds', {})
 
     for cloud in user_defined_clouds:
-        CLOUDS.update(cloud)
+        DEFAULT_CLOUDS.update(cloud)
 
-    supported_clouds = list(CLOUDS.keys())
+    supported_clouds = list(DEFAULT_CLOUDS.keys())
     selected = prompt_choice_list('Select a cloud', supported_clouds)
 
-    profile['cloud'] = CLOUDS.get(supported_clouds[selected])
+    profile['cloud'] = DEFAULT_CLOUDS.get(supported_clouds[selected])
     write_profile(json.dumps(profile))
 
     print(f'Selected {supported_clouds[selected]}')
-    # Save selected cloud in profile.json
 
 
 def add_cloud(name: str, endpoint: str, authority: str):
+    #TODO: validate that endpoint and authority are valid urls
+
     cloud = {name: {'endpoint': endpoint, 'authority': authority}}
 
     profile = read_profile()
@@ -47,27 +41,3 @@ def add_cloud(name: str, endpoint: str, authority: str):
 
     write_profile(json.dumps(profile))
     print(f'Cloud "{name}" added successfully')
-
-
-def reset_cloud():
-    print('Resetting a cloud')
-
-
-#TODO: Abstract reading and writing to profile
-def read_profile() -> dict:
-    result = None
-
-    try:
-        with open(PROFILE_LOCATION, 'r') as file:
-            result = file.read()
-        return json.loads(result)
-    except IOError as ex:
-        return {}
-
-
-def write_profile(profile: str):
-    try:
-        with open(PROFILE_LOCATION, 'w') as file:
-            file.write(profile)
-    except IOError as ex:
-        raise CLIException('Selected cloud was not set, CLI will use the PublicCLoud') from ex
